+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Only recommend examples on restricted strings.
+- Add the possibility to include a URL for further reference to rule sets, which is displayed in the output.
+
+## [1.1.0] - 2023-04-05
+
 - Add rule to check whether root-level specifies properties of the common schema structure.
 - Add rule to check that objects are non-empty.
 
@@ -93,7 +97,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added first basic linting.
 
-[Unreleased]: https://github.com/giantswarm/schemalint/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/giantswarm/schemalint/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/giantswarm/schemalint/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/giantswarm/schemalint/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/giantswarm/schemalint/compare/v0.10.0...v1.0.0
 [0.10.0]: https://github.com/giantswarm/schemalint/compare/v0.9.0...v0.10.0